+++ languages/en/tl_member.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * Member contact settings extension for Contao Open Source CMS.
+ *
+ * @copyright  Copyright (c) 2017 Contao Bayern
+ * @author     Jörg Moldenhauer
+ * @author     Andreas Fieger
+ * @license    https://opensource.org/licenses/lgpl-3.0.html LGPL-3.0
+ *
+ * @see        https://github.com/ContaoBayern/contao-member-contact-settings
+ */
+
+// Legends
+$GLOBALS['TL_LANG']['tl_member']['contactSettings_legend'] = 'Contact-settings';
+$GLOBALS['TL_LANG']['tl_member']['contactSettingsDetails'] = 'Contact-settings';
+
+// Fields
+$GLOBALS['TL_LANG']['tl_member']['contactEmail'] = ['Contact by e-mail', 'I want to be contacted by e-mail.'];
+$GLOBALS['TL_LANG']['tl_member']['contactLetter'] = ['Contact by letter', 'I want to be contacted by letter.'];
+$GLOBALS['TL_LANG']['tl_member']['contactPhone'] = ['Contact by phone', 'I want to be contacted by phone.'];
+$GLOBALS['TL_LANG']['tl_member']['contactFax'] = ['Contact yb fax', 'I want to be contacted by fax.'];