+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Contao Member Contact Settings change log
 
+
+## 1.0.0
+
+### New
+* Grouped and ungrouped personal data templates are supported (see #13)
+* English translation is added (see #21)
+
+### Changed
+* A hook is used to add "novalidate" attribute to member data form instead of using an own template (see #14)
+
+### Fixed
+* Fixed Bug which prohibited deactivation of saved contact settings in member data frontend module (see #22)
+
+
 ## 0.2.1
 
 ### New